+++ server/seeds/01_organisation-developers.js
@@ -46,6 +46,7 @@ exports.seed = async knex => {
     await knex('user_organisation').insert([
         { organisation_id: organisations[0].id, user_id: users[2].id }, //viktor -> dhis2
         { organisation_id: organisations[1].id, user_id: users[1].id }, //erik -> who
+        { organisation_id: organisations[0].id, user_id: users[1].id }, //erik -> dhis2, we need a user with multiple organisations to test that this also works
     ])
 
     console.log('Inserting user-organisations #02')