+++ server/src/models/v2/User.js
@@ -1,13 +1,13 @@
 const joi = require('@hapi/joi')
-const {
-    definition: defaultDefinition,
-    createDefaultValidator,
-} = require('./Default')
+const { definition: defaultDefinition } = require('./Default')
+const { createDefaultValidator } = require('./helpers')
 
-const definition = defaultDefinition.append({
-    name: joi.string(),
-    email: joi.string(),
-})
+const definition = defaultDefinition
+    .append({
+        name: joi.string(),
+        email: joi.string(),
+    })
+    .label('User')
 
 const parseDatabaseJson = createDefaultValidator(definition)