+++ server/src/plugins/errorMapper.js
@@ -11,6 +11,12 @@ const {
     wrapError,
 } = require('db-errors')
 
+/**
+ * A plugin that parses database-errors thrown in handlers and maps them to the correct boom-error.
+ * This means that in many cases we do not need to handle or catch db-query errors
+ * unless we need to do something else with the error.
+ */
+
 const dbErrorMap = {
     badRequest: [
         CheckViolationError,
@@ -27,7 +33,7 @@ const dbErrorMap = {
 const onPreResponseHandler = function(request, h) {
     const { response: error } = request
     // not error or joi-error - ignore
-    // By default validation errors thrown in handler are transformed to 500-error, as they are not Boom-errors.
+    // By default validation errors thrown in har are transformed to 500-error, as they are not Boom-errors.
     // This makes sense, as internal validation errors (outside validation-handlers) are developer errors
     if (!error.isBoom || error.isJoi) {
         return h.continue